Tests fail on Windows due to spurious CR characters
Given that both the tests and Module::Install have been around
for a while, my explanation for the failures is that the commit

https://github.com/Perl-Toolchain-Gang/ExtUtils-MakeMaker/commit/cfabc28

to ExtUtils::MakeMaker is causing the problem.

Coming from that perspective, I conditionally applied binmode
':raw' to filehandles in t/lib/Test.pm. Tests in t/22_installdirs.t
were still failing after that, so I set the captures up to any CR/LF
characters.

This seems to work for now, although I have yet to test it elsehwere.