remove dependency on '.' being in @INC
Perl 5.25.7 has the option of not including '.' in @INC by default.  This should rework M::I so
that its tests pass under this configuration.  t/lib is added to @INC for the test suite where
needed, and Test.pm has been renamed to MyTest.pm so that it is not confused with the Test.pm
that comes with Perl.