Add URLs for licences listed in CPAN::Meta::Spec
This is to fix licence URL in metadata for licenses only listed in
CPAN::Meta::Spec like "gpl_2", etc.

It's true that we already have "gpl2" in %license_urls, sadly that we
can't write "license 'gpl2';" in Makefile.PL because the license name is
validated by CPAN::Meta where "gpl2" is invalid.